# PaperLess.Frontend/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Mock Server",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/tags" : {
      "get" : {
        "operationId" : "GetTags",
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Tags" ]
      },
      "post" : {
        "operationId" : "CreateTag",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewTag"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewTag"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewTag"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Tags" ]
      }
    },
    "/api/tags/{id}" : {
      "delete" : {
        "operationId" : "DeleteTag",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Tags" ]
      },
      "put" : {
        "operationId" : "UpdateTag",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Tag"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Tag"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Tag"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Tags" ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "NewTag" : {
        "additionalProperties" : false,
        "example" : {
          "matching_algorithm" : 0,
          "color" : "color",
          "is_insensitive" : true,
          "name" : "name",
          "match" : "match",
          "is_inbox_tag" : true
        },
        "properties" : {
          "name" : {
            "nullable" : true,
            "type" : "string"
          },
          "color" : {
            "nullable" : true,
            "type" : "string"
          },
          "match" : {
            "nullable" : true,
            "type" : "string"
          },
          "matching_algorithm" : {
            "format" : "int64",
            "type" : "integer"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "is_inbox_tag" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "Tag" : {
        "additionalProperties" : false,
        "example" : {
          "matching_algorithm" : 6,
          "document_count" : 1,
          "color" : "color",
          "is_insensitive" : true,
          "name" : "name",
          "match" : "match",
          "id" : 0,
          "is_inbox_tag" : true,
          "slug" : "slug"
        },
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "slug" : {
            "nullable" : true,
            "type" : "string"
          },
          "name" : {
            "nullable" : true,
            "type" : "string"
          },
          "color" : {
            "nullable" : true,
            "type" : "string"
          },
          "match" : {
            "nullable" : true,
            "type" : "string"
          },
          "matching_algorithm" : {
            "format" : "int64",
            "type" : "integer"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "is_inbox_tag" : {
            "type" : "boolean"
          },
          "document_count" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    }
  }
}